app/src/test/java/com/github/notizklotz/derbunddownloader/common/DateHandlingUtilsTest.java
/*
 * Der Bund ePaper Downloader - App to download ePaper issues of the Der Bund newspaper
 * Copyright (C) 2013 Adrian Gygax
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
 * GNU General Public License for more details.
 *
 * You should have received a copy of the GNU General Public License
 * along with this program. If not, see {http://www.gnu.org/licenses/}.
 */

package com.github.notizklotz.derbunddownloader.common;

import android.test.suitebuilder.annotation.SmallTest;

import org.joda.time.DateTime;
import org.joda.time.DateTimeZone;
import org.junit.Test;

import static org.junit.Assert.assertEquals;

@SmallTest
public class DateHandlingUtilsTest {

    @Test
    public void testToFullStringDefaultTimezone() {
        assertEquals("31.12.1969 19:00:00 EST", DateHandlingUtils.toFullString(new DateTime(0), DateTimeZone.forID("America/New_York")));
    }

}
